button ambiguation achieved
now the player has 4 buttons, the action button is always the main button to move forward, and can have a consistant button bind n such. but the state machine is now monitoring for button presses from players and doing the state based actions that are needed from all the other parts of the system , as the game manager should be.